Refactor icu_timezone string parsing
