+++ components/timezone/src/types.rs
@@ -75,14 +75,11 @@ impl GmtOffset {
     /// ```
     /// use icu::timezone::GmtOffset;
     ///
-    /// let offset0: GmtOffset =
-    ///     GmtOffset::try_from_str("Z").expect("Failed to parse a time zone");
-    /// let offset1: GmtOffset =
-    ///     GmtOffset::try_from_str("+05").expect("Failed to parse a time zone");
-    /// let offset2: GmtOffset = GmtOffset::try_from_str("+0500")
-    ///     .expect("Failed to parse a time zone");
-    /// let offset3: GmtOffset = GmtOffset::try_from_str("-05:00")
-    ///     .expect("Failed to parse a time zone");
+    /// let offset0: GmtOffset = GmtOffset::try_from_str("Z").unwrap();
+    /// let offset1: GmtOffset = GmtOffset::try_from_str("+05").unwrap();
+    /// let offset2: GmtOffset = GmtOffset::try_from_str("+0500").unwrap();
+    /// let offset3: GmtOffset = GmtOffset::try_from_str("-05:00").unwrap();
+    ///
     /// let offset_err0 =
     ///     GmtOffset::try_from_str("0500").expect_err("Invalid input");
     /// let offset_err1 =
